v0.7.1 Enable message chunking
There was a problem when some messages were not being received.  This
was due to a message size limitation in the discord api.

2000 char limit for normal users
4000 char limit for nitro users

A chunk size of 1980 was set, this is to allow "wrapper" chars,
typically used to designate code blocks, to be added.

The chunking tries to do it cleanly on end of line boundaries to make
the messages more readable.